pyproject.toml
[project]
name = "llm-book"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.6",
    "langchain>=0.1,<0.3",
    "langchain-openai>=0.1",
    "langchain-core>=0.1",
    "langchain-community>=0.0.1",
    "langgraph>=0.0.20",
    "python-dotenv>=1.0",
    "pydantic>=2.5",
    "yfinance>=0.2.28",
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "feedparser>=6.0",
    "html2text>=2020.1.16",
    "faiss-cpu",
    "rank-bm25",
    "numpy",
    "pandas",
    "tiktoken",
]
